refactor: re-arrange tests and extract error messages and strings
