+++ example/src/tests/unit/common.ts
@@ -6,6 +6,14 @@ import {
 import { resetTestDb } from '../db'
 import chai from 'chai'
 
+export const TEST_ERROR_CODES = {
+  EXPECT_NITRO_SQLITE_ERROR: 'Should have thrown a valid NitroSQLiteError',
+  EXPECT_PROMISE_REJECTION: 'Should have thrown a promise rejection',
+} as const
+
+export const TEST_ERROR_MESSAGE = 'Error from callback'
+export const TEST_ERROR = new Error(TEST_ERROR_MESSAGE)
+
 export function isNitroSQLiteError(e: unknown): e is NitroSQLiteError {
   return e instanceof NitroSQLiteError
 }
