+++ example/src/tests/unit/index.ts
@@ -1,9 +1,10 @@
 import { beforeEach, describe } from '../MochaRNAdapter'
 import { setupTestDb } from './common'
-import registerExecuteUnitTests from './specs/execute.spec'
-import registerTransactionUnitTests from './specs/transaction.spec'
-import registerExecuteBatchUnitTests from './specs/executeBatch.spec'
+import registerExecuteUnitTests from './specs/operations/execute.spec'
+import registerTransactionUnitTests from './specs/operations/transaction.spec'
+import registerExecuteBatchUnitTests from './specs/operations/executeBatch.spec'
 import registerTypeORMUnitTestsSpecs from './specs/typeorm.spec'
+import registerDatabaseQueueUnitTests from './specs/DatabaseQueue.spec'
 
 export function registerUnitTests() {
   beforeEach(setupTestDb)
@@ -13,6 +14,8 @@ export function registerUnitTests() {
     registerTransactionUnitTests()
     registerExecuteBatchUnitTests()
   })
+
+  registerDatabaseQueueUnitTests()
 }
 
 export function registerTypeORMUnitTests() {
